+++ .github/workflows/windows-x86.yml
@@ -1,0 +1,51 @@
+name: Windows x86
+
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+    branches:
+      - master
+
+env:
+  CTEST_OUTPUT_ON_FAILURE: 1
+
+jobs:
+  build:
+    runs-on: windows-latest
+
+    steps:
+      - uses: actions/checkout@v1
+
+      - name: configure
+        run: cmake -Htest -Bbuild/test -DCMAKE_GENERATOR_PLATFORM=Win32
+
+      - name: build
+        run: cmake --build build/test --config Debug -j4
+
+      - name: test
+        run: |
+          cd build/test
+          ctest --build-config Debug
+          cd ../../
+          del /S /Q /F .\build
+        shell: cmd
+
+      - name: install php
+        run: |
+          .\.github/workflows/bin/install-php7.4-win-x86.ps1
+
+      - name: build and install library with php
+        run: |
+          $phpDevPath=$([environment]::GetEnvironmentvariable("PHP_DEV_PATH", "Machine"))
+          cmake -H"." -Bbuild -DCMAKE_BUILD_TYPE=Release -DCMAKE_GENERATOR_PLATFORM=Win32 -DWITH_PHP=1 -DPHP_INCLUDE_DIR="$phpDevPath\include" -DPHP_LIB_DIR="$phpDevPath\lib"
+          cmake --build build --target install --config Release -j
+
+      - name: php test
+        run: |
+          [environment]::SetEnvironmentvariable("PATH", $([environment]::GetEnvironmentvariable("PATH", "Machine")))
+          cd phptest
+          composer update
+          phpunit
+          cd ../
