+++ .travis/php.ini
@@ -1,0 +1,1 @@
+extension=ffi
