+++ README.md
@@ -1,22 +1,12 @@
 # ChineseUtil
 
-PHP 中文工具类，支持汉字转拼音、拼音分词、简繁互转。
-
-PHP Chinese Tool class, support Chinese pinyin, pinyin participle, simplified and traditional conversion
-
-目前本类库拥有的三个功能，都是在实际开发过程中整理出来的。这次使用的数据不同于以前我开源过汉字转拼音和简繁互转，数据都是从字典网站采集下来的，比以前的数据更加准确。
+PHP 中文工具类，支持汉字转拼音、拼音分词、简繁互转、数字转换、金额数字转换。
 
 由于中文的博大精深，字有多音字，简体字和繁体字也有多种对应。并且本类库返回的所有结果，均为包含所有组合的数组。
 
 本类库字典数据总共收录 73925 个汉字，包括：3955 个简体字，1761 个繁体字，68209 个其它汉字。
 
-## 内存占用
-
-类库第一个版本发布开始，群里朋友就展开了激烈的讨论，最大的问题就在于内存占用以及性能问题上。经过我不断尝试几种方案，最终决定设置三种模式，来适应不同用户之间的需求。
-
-我已经尝试过 Redis，速度比 SQLite 慢 3 倍，没有必要加入支持。所以目前来看，SQLite 模式性价比最高！
-
----
+## 模式
 
 ### 性能模式 (Memory)
 
@@ -44,6 +34,18 @@ PHP Chinese Tool class, support Chinese pinyin, pinyin participle, simplified an
 
 由于精简了数据，一些拼音结果需要经过代码计算处理才可以得出，所以性能较差。
 
+### FFI 模式 (FFI)
+
+需要 PHP >= 7.4 并且启用 FFI 扩展，代码全部由 C++ 开发，性能和内存占用都比 PHP 实现的要好。
+
+> FFI 动态库 C++ 代码：<https://github.com/Yurunsoft/chinese-util-cpp>
+
+### Swoole FFI 模式 (SwooleFFI)
+
+需要 PHP >= 7.4 并且启用 FFI、Swoole 扩展，代码全部由 C++ 开发，性能和内存占用都比 PHP 实现的要好。
+
+不会阻塞 PHP 代码所在线程。
+
 ---
 
 默认情况下，优先使用通用模式，如果环境不支持 PDO 将采用兼容模式。
@@ -57,6 +59,10 @@ Chinese::setMode('Memory');
 Chinese::setMode('SQLite');
 // 设为兼容模式
 Chinese::setMode('JSON');
+// 设为 FFI 模式
+Chinese::setMode('FFI');
+// 设为Swoole FFI 模式
+Chinese::setMode('SwooleFFI');
 ```
 
 无论何种模式，拼音分词所需数据总是从 JSON 数据中加载。
