+++ clib/include.h
@@ -1,0 +1,1 @@
+void init_chinese_util();
