--- data/pinyinSound.json
@@ -1,1 +1,0 @@
-{"a":{"ab":"a","tone":0},"ā":{"ab":"a","tone":1},"á":{"ab":"a","tone":2},"ǎ":{"ab":"a","tone":3},"à":{"ab":"a","tone":4},"o":{"ab":"o","tone":0},"ō":{"ab":"o","tone":1},"ó":{"ab":"o","tone":2},"ǒ":{"ab":"o","tone":3},"ò":{"ab":"o","tone":4},"e":{"ab":"e","tone":0},"ē":{"ab":"e","tone":1},"é":{"ab":"e","tone":2},"ě":{"ab":"e","tone":3},"è":{"ab":"e","tone":4},"ī":{"ab":"i","tone":1},"í":{"ab":"i","tone":2},"ǐ":{"ab":"i","tone":3},"ì":{"ab":"i","tone":4},"ū":{"ab":"u","tone":1},"ú":{"ab":"u","tone":2},"ǔ":{"ab":"u","tone":3},"ù":{"ab":"u","tone":4},"ü":{"ab":"v","tone":0},"ǖ":{"ab":"v","tone":1},"ǘ":{"ab":"v","tone":2},"ǚ":{"ab":"v","tone":3},"ǜ":{"ab":"v","tone":4}}
