--- data/pinyinSplitData.json
@@ -1,1 +1,0 @@
-{"shengmu":["a","o","e","b","p","m","f","d","t","n","l","g","k","h","j","q","x","r","z","c","s","y","w","zh","ch","sh","er","a","o","e","ai","ei","ao","ou"],"yunmu":["i","u","v","a","o","e","ia","ua","uo","ie","ue","ai","ei","ao","ou","an","en","in","un","ui","uai","iao","uan","ang","eng","ing","ong","iong","iang","uang"],"relation":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"b":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true,"a":{"n":{"py":true},"o":{"py":true}},"e":{"py":true},"n":{"py":true,"g":{"py":true}}},"o":{"py":true},"u":{"py":true,"n":{"py":true}}},"c":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"n":{"py":true,"g":{"py":true}}},"h":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"i":{"py":true},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"d":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true,"a":{"py":true,"n":{"py":true},"o":{"py":true}},"e":{"py":true},"n":{"g":{"py":true}},"u":{"py":true}},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"r":{"py":true}},"f":{"a":{"py":true,"n":{"py":true,"g":{"py":true}}},"e":{"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"a":{"o":{"py":true}}},"o":{"py":true,"u":{"py":true}},"u":{"py":true}},"g":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"h":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"j":{"i":{"py":true,"a":{"py":true,"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"n":{"g":{"py":true}}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"e":{"py":true},"n":{"py":true}}},"k":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"l":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true,"a":{"py":true,"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true},"n":{"py":true,"g":{"py":true}},"u":{"py":true}},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"e":{"py":true},"n":{"py":true},"o":{"py":true}},"v":{"py":true}},"m":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true},"s":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true,"a":{"n":{"py":true},"o":{"py":true}},"e":{"py":true},"n":{"py":true,"g":{"py":true}},"u":{"py":true}},"o":{"py":true,"u":{"py":true}},"u":{"py":true}},"n":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true,"a":{"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true},"n":{"py":true,"g":{"py":true}},"u":{"py":true}},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"e":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"v":{"py":true,"e":{"py":true}}},"o":{"py":true,"e":{"s":{"py":true}},"u":{"py":true}},"p":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"l":{"py":true}}},"h":{"a":{"s":{"py":true}},"d":{"e":{"n":{"g":{"py":true}}}},"o":{"s":{"py":true}}},"i":{"py":true,"a":{"n":{"py":true},"o":{"py":true}},"e":{"py":true},"n":{"py":true,"g":{"py":true}}},"o":{"py":true,"u":{"py":true}},"p":{"u":{"n":{"py":true}}},"u":{"py":true}},"q":{"i":{"py":true,"a":{"py":true,"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"n":{"g":{"py":true}}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"e":{"py":true},"n":{"py":true}}},"r":{"a":{"py":true,"m":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"py":true,"n":{"py":true}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"s":{"a":{"py":true,"e":{"n":{"g":{"py":true}}},"i":{"py":true},"l":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"n":{"py":true}}},"h":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"o":{"u":{"py":true}},"u":{"py":true,"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"i":{"py":true},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"t":{"a":{"py":true,"e":{"py":true},"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"g":{"py":true}},"u":{"l":{"py":true}}},"i":{"py":true,"a":{"n":{"py":true},"o":{"py":true}},"e":{"py":true},"n":{"g":{"py":true}}},"o":{"n":{"py":true,"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"u":{"u":{"py":true}},"w":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"e":{"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"o":{"py":true},"u":{"py":true}},"x":{"i":{"py":true,"a":{"py":true,"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"n":{"g":{"py":true}}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"e":{"py":true},"n":{"py":true}}},"y":{"a":{"py":true,"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true},"i":{"py":true,"n":{"py":true,"g":{"py":true}}},"o":{"py":true,"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"e":{"py":true},"n":{"py":true}}},"z":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"h":{"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}},"o":{"py":true}},"e":{"py":true,"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"py":true,"i":{"py":true},"n":{"py":true,"g":{"py":true}}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}},"i":{"py":true},"o":{"n":{"g":{"py":true}},"u":{"py":true}},"u":{"py":true,"a":{"n":{"py":true}},"i":{"py":true},"n":{"py":true},"o":{"py":true}}}}}
