+++ src/Chinese/Driver/SimplifiedTraditional/SwooleFFI.php
@@ -1,0 +1,33 @@
+<?php
+namespace Yurun\Util\Chinese\Driver\SimplifiedTraditional;
+
+use Yurun\Util\Chinese\FFIDriver;
+
+class SwooleFFI implements BaseInterface
+{
+    public function __construct()
+    {
+        FFIDriver::getHandler();
+    }
+
+    /**
+     * 繁体转简体
+     * @param string $string
+     * @return array
+     */
+    public function toSimplified($string)
+    {
+        return swoole_convert_to_simplified($string);
+    }
+
+    /**
+     * 简体转繁体
+     * @param string $string
+     * @return array
+     */
+    public function toTraditional($string)
+    {
+        return swoole_convert_to_traditional($string);
+    }
+
+}
