+++ src/Chinese/Traits/JSONInit.php
@@ -34,7 +34,7 @@ trait JSONInit
             }
             else if(empty(Chinese::$option['pinyinSoundDataPath']))
             {
-                Chinese::$chineseData['pinyinSound'] = json_decode(file_get_contents(dirname(dirname(dirname(__DIR__))) . '/data/pinyinSound.json'), true);
+                Chinese::$chineseData['pinyinSound'] = json_decode(file_get_contents(dirname(dirname(dirname(__DIR__))) . '/data/pinyinData.json'), true)['sound'];
             }
             else
             {
