+++ tests/.phpunit.result.cache
@@ -1,0 +1,1 @@
+C:37:"PHPUnit\Runner\DefaultTestResultCache":5873:{a:2:{s:7:"defects";a:0:{}s:5:"times";a:78:{s:49:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testMode";d:0.018;s:52:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testPinyin1";d:0.404;s:52:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testPinyin2";d:0;s:54:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testPinyinAll";d:0;s:56:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testPinyinFirst";d:0;s:56:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testPinyinSound";d:0;s:62:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testPinyinSoundNumber";d:0;s:57:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testPinyinCustom";d:0;s:64:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testPinyinSplitNoPinyin";d:0;s:67:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testPinyinNotSplitNoPinyin";d:0;s:56:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testPinyinSplit";d:0.002;s:69:"Yurun\Util\ChineseUtil\Test\FFIModeTest::testSimplifiedAndTraditional";d:0.001;s:50:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testMode";d:0;s:53:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testPinyin1";d:0.137;s:53:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testPinyin2";d:0;s:55:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testPinyinAll";d:0;s:57:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testPinyinFirst";d:0;s:57:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testPinyinSound";d:0;s:63:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testPinyinSoundNumber";d:0;s:58:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testPinyinCustom";d:0;s:65:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testPinyinSplitNoPinyin";d:0;s:68:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testPinyinNotSplitNoPinyin";d:0;s:57:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testPinyinSplit";d:0.002;s:70:"Yurun\Util\ChineseUtil\Test\JSONModeTest::testSimplifiedAndTraditional";d:0.002;s:52:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testMode";d:0.005;s:55:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testPinyin1";d:0.462;s:55:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testPinyin2";d:0;s:57:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testPinyinAll";d:0;s:59:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testPinyinFirst";d:0;s:59:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testPinyinSound";d:0;s:65:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testPinyinSoundNumber";d:0;s:60:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testPinyinCustom";d:0;s:67:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testPinyinSplitNoPinyin";d:0;s:70:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testPinyinNotSplitNoPinyin";d:0;s:59:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testPinyinSplit";d:0.001;s:72:"Yurun\Util\ChineseUtil\Test\MemoryModeTest::testSimplifiedAndTraditional";d:0.001;s:56:"Yurun\Util\ChineseUtil\Test\Money\FFIMoneyTest::testMode";d:0.018;s:61:"Yurun\Util\ChineseUtil\Test\Money\FFIMoneyTest::testToChinese";d:0.002;s:60:"Yurun\Util\ChineseUtil\Test\Money\FFIMoneyTest::testToNumber";d:0.002;s:59:"Yurun\Util\ChineseUtil\Test\Money\MemoryMoneyTest::testMode";d:0;s:64:"Yurun\Util\ChineseUtil\Test\Money\MemoryMoneyTest::testToChinese";d:0.002;s:63:"Yurun\Util\ChineseUtil\Test\Money\MemoryMoneyTest::testToNumber";d:0;s:62:"Yurun\Util\ChineseUtil\Test\Money\SwooleFFIMoneyTest::testMode";d:0;s:67:"Yurun\Util\ChineseUtil\Test\Money\SwooleFFIMoneyTest::testToChinese";d:0.001;s:66:"Yurun\Util\ChineseUtil\Test\Money\SwooleFFIMoneyTest::testToNumber";d:0;s:58:"Yurun\Util\ChineseUtil\Test\Number\FFINumberTest::testMode";d:0;s:63:"Yurun\Util\ChineseUtil\Test\Number\FFINumberTest::testToChinese";d:0.002;s:62:"Yurun\Util\ChineseUtil\Test\Number\FFINumberTest::testToNumber";d:0;s:61:"Yurun\Util\ChineseUtil\Test\Number\MemoryNumberTest::testMode";d:0;s:66:"Yurun\Util\ChineseUtil\Test\Number\MemoryNumberTest::testToChinese";d:0.001;s:65:"Yurun\Util\ChineseUtil\Test\Number\MemoryNumberTest::testToNumber";d:0;s:64:"Yurun\Util\ChineseUtil\Test\Number\SwooleFFINumberTest::testMode";d:0;s:69:"Yurun\Util\ChineseUtil\Test\Number\SwooleFFINumberTest::testToChinese";d:0.001;s:68:"Yurun\Util\ChineseUtil\Test\Number\SwooleFFINumberTest::testToNumber";d:0;s:52:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testMode";d:0;s:55:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testPinyin1";d:0.003;s:55:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testPinyin2";d:0.001;s:57:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testPinyinAll";d:0.002;s:59:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testPinyinFirst";d:0.001;s:59:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testPinyinSound";d:0.001;s:65:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testPinyinSoundNumber";d:0.001;s:60:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testPinyinCustom";d:0.001;s:67:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testPinyinSplitNoPinyin";d:0.001;s:70:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testPinyinNotSplitNoPinyin";d:0.001;s:59:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testPinyinSplit";d:0.001;s:72:"Yurun\Util\ChineseUtil\Test\SQLiteModeTest::testSimplifiedAndTraditional";d:0.005;s:55:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testMode";d:0;s:58:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testPinyin1";d:0.001;s:58:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testPinyin2";d:0;s:60:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testPinyinAll";d:0;s:62:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testPinyinFirst";d:0;s:62:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testPinyinSound";d:0;s:68:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testPinyinSoundNumber";d:0;s:63:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testPinyinCustom";d:0;s:70:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testPinyinSplitNoPinyin";d:0;s:73:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testPinyinNotSplitNoPinyin";d:0;s:62:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testPinyinSplit";d:0.001;s:75:"Yurun\Util\ChineseUtil\Test\SwooleFFIModeTest::testSimplifiedAndTraditional";d:0.001;}}}
