+++ tests/unit/Money/FFINumberTest.php
@@ -1,0 +1,16 @@
+<?php
+namespace Yurun\Util\ChineseUtil\Test\Money;
+
+/**
+ * @testdox FFI Mode 中文金额转换
+ */
+class FFIMoneyTest extends BaseMoneyTest
+{
+    /**
+     * 模式
+     *
+     * @var string
+     */
+    protected $mode = 'FFI';
+
+}
