+++ tests/unit/Money/MemoryNumberTest.php
@@ -1,0 +1,16 @@
+<?php
+namespace Yurun\Util\ChineseUtil\Test\Money;
+
+/**
+ * @testdox Memory Mode 中文金额转换
+ */
+class MemoryMoneyTest extends BaseMoneyTest
+{
+    /**
+     * 模式
+     *
+     * @var string
+     */
+    protected $mode = 'Memory';
+
+}
