+++ tests/unit/Number/BaseNumberTest.php
@@ -1,16 +1,37 @@
 <?php
-namespace Yurun\Util\ChineseUtil\Test;
+namespace Yurun\Util\ChineseUtil\Test\Number;
 
+use Yurun\Util\Chinese;
 use Yurun\Util\Chinese\Number;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @testdox 中文数字转换
  */
-class NumberTest extends TestCase
+abstract class BaseNumberTest extends TestCase
 {
+    /**
+     * 模式
+     *
+     * @var string
+     */
+    protected $mode;
+
+    protected function check()
+    {
+
+    }
+
+    public function testMode()
+    {
+        $this->check();
+        Chinese::setMode($this->mode);
+        $this->assertEquals($this->mode, Chinese::getMode());
+    }
+
     public function testToChinese()
     {
+        $this->check();
         // 数字
         $this->assertEquals('五', Number::toChinese(5));
         $this->assertEquals('一十二', Number::toChinese(12));
@@ -33,6 +54,7 @@ class NumberTest extends TestCase
 
     public function testToNumber()
     {
+        $this->check();
         // 数字
         $this->assertEquals(5, Number::toNumber('五'));
         $this->assertEquals(12, Number::toNumber('一十二'));
