+++ tests/unit/Number/FFINumberTest.php
@@ -1,0 +1,16 @@
+<?php
+namespace Yurun\Util\ChineseUtil\Test\Number;
+
+/**
+ * @testdox FFI Mode 中文数字转换
+ */
+class FFINumberTest extends BaseNumberTest
+{
+    /**
+     * 模式
+     *
+     * @var string
+     */
+    protected $mode = 'FFI';
+
+}
