+++ tests/unit/Number/MemoryNumberTest.php
@@ -1,0 +1,16 @@
+<?php
+namespace Yurun\Util\ChineseUtil\Test\Number;
+
+/**
+ * @testdox Memory Mode 中文数字转换
+ */
+class MemoryNumberTest extends BaseNumberTest
+{
+    /**
+     * 模式
+     *
+     * @var string
+     */
+    protected $mode = 'Memory';
+
+}
